requirements.txt
coverage==3.7.1
Django==1.8.3
django-markdown-deux==1.0.5
django-pagedown==0.1.0
factory-boy==2.5.2
gunicorn==19.3.0
markdown2==2.3.0
Pillow==2.9.0
Pygments==2.0.2
wheel==0.24.0
